+++ main.py
@@ -1,3 +1,10 @@
 import vsdownload.vsdownload
 
 vsdownload.vsdownload.console_script()
+
+# some test commands (root)
+# python main.py --help
+# python main.py save https://multiplatform-f.akamaihd.net/i/multi/will/bunny/big_buck_bunny_,640x360_400,640x360_700,640x360_1000,950x540_1500,.f4v.csmil/index_0_av.m3u8
+
+# resumable downloading
+# python main.py save https://multiplatform-f.akamaihd.net/i/multi/will/bunny/big_buck_bunny_,640x360_400,640x360_700,640x360_1000,950x540_1500,.f4v.csmil/index_0_av.m3u8 --no-cleanup